I added a few new features: 1. Ignore SQL DataSync - This defaults to true. When selected the tables and procedures SQL Server adds when using DataSync adds a lot of data.
My day job recently did this and 57 tables and a couple of hundred views added 717 tables / procedures.

I also now ignore database_firewall_rules view that azure adds by default, but there is a checkbox so you can change this value.

And finally I added code to run a comparison async, and it helped with the app no longer says (Not Responding), but the mouse cursor is still an hour glass so perhaps more can be done.

The method call to DataJuggler.Net.LoadDatabaseSchema might need to also be async for this to improve the UI responsiveness more while a comparision is running.

Also DataJuggler.Net Nuget Package has been updated to version 5.5.0 to include these changes.